+++ CHANGES.md
@@ -10,6 +10,8 @@
 
 * FfprobeCharacterization helper class for a/v characterization metadata https://github.com/sciencehistory/kithe/pull/139
 
+* Create an Asset#file_metadata method that delegates to file.metadata for convenience https://github.com/sciencehistory/kithe/pull/140
+
 *
 
 ### Changed