+++ CHANGES.md
@@ -8,7 +8,7 @@
 
 ### Added
 
-*
+* Tools for using a standard shrine derivatives processor with kithe features, including lifecycle management and guard options. https://github.com/sciencehistory/kithe/pull/143
 
 *