+++ CHANGES.md
@@ -8,13 +8,21 @@
 
 ### Added
 
-* Tools for using a standard shrine derivatives processor with kithe features, including lifecycle management and guard options. https://github.com/sciencehistory/kithe/pull/143
+*
 
-* add Kithe::FfmpegExtractJpg service for extracting frame from video  https://github.com/sciencehistory/kithe/pull/144
+*
 
 ### Changed
 
 *
+
+## 2.5.0 (2 Mar 2022)
+
+### Added
+
+* Tools for using a standard shrine derivatives processor with kithe features, including lifecycle management and guard options. https://github.com/sciencehistory/kithe/pull/143
+
+* add Kithe::FfmpegExtractJpg service for extracting frame from video  https://github.com/sciencehistory/kithe/pull/144
 
 
 ## 2.4.0 (14 Feb 2022)